refactor(platform): split cache.rs into cache module (Task 5.6)
